[libc] Change exit status from int to long
Slightly breaking the C standard, but in a way that's unlikely to break
things - allow 64-bit process exit status codes.

+++ src/kernel/objects/process.cpp
@@ -46,7 +46,7 @@ process::create_kernel_process(page_table *pml4)
 }
 
 void
-process::exit(int32_t code)
+process::exit(int64_t code)
 {
     if (m_state == state::exited)
         return;
